fix drop area border, rename FileDrop to FileSelect
