--- src/shared/components/FileDrop/index.ts
@@ -1,3 +1,0 @@
-import FileDrop from './FileDrop'
-
-export default FileDrop
