+++ src/shared/components/FileSelect/index.ts
@@ -1,0 +1,3 @@
+import FileSelect from './FileSelect'
+
+export default FileSelect
