+++ src/shared/components/index.ts
@@ -32,5 +32,5 @@ export { default as Checkbox } from './Checkbox'
 export { default as Textarea } from './TextArea'
 export { default as ActionBar } from './ActionBar'
 export { default as MultiFileSelect } from './MultiFileSelect'
-export { default as FileDrop } from './FileDrop'
+export { default as FileSelect } from './FileSelect'
 export { default as FileStep } from './FileStep'
